+++ ashley-extensions/src/com/roaringcatgames/kitten2d/ashley/systems/FollowerSystem.java
@@ -2,9 +2,13 @@ package com.roaringcatgames.kitten2d.ashley.systems;
 
 import com.badlogic.ashley.core.*;
 import com.badlogic.ashley.systems.IteratingSystem;
+import com.badlogic.gdx.math.MathUtils;
 import com.badlogic.gdx.math.Vector2;
+import com.badlogic.gdx.math.Vector3;
 import com.badlogic.gdx.utils.Array;
+import com.roaringcatgames.kitten2d.ashley.K2ComponentMappers;
 import com.roaringcatgames.kitten2d.ashley.VectorUtils;
+import com.roaringcatgames.kitten2d.ashley.components.FollowMode;
 import com.roaringcatgames.kitten2d.ashley.components.FollowerComponent;
 import com.roaringcatgames.kitten2d.ashley.components.TransformComponent;
 
@@ -15,6 +19,8 @@ public class FollowerSystem extends IteratingSystem {
 
     ComponentMapper<FollowerComponent> fm;
     ComponentMapper<TransformComponent> tm;
+
+    private Vector2 moveToAdjustment = new Vector2();
 
     Array<Entity> queue;
     Family familyToWatchForRemovals;
@@ -69,27 +75,72 @@ public class FollowerSystem extends IteratingSystem {
 
     @Override
     protected void processEntity(Entity entity, float deltaTime) {
+
         FollowerComponent fc = fm.get(entity);
         if(fc.target != null) {
             TransformComponent targetPos = tm.get(fc.target);
             if (targetPos != null) {
+                TransformComponent tc = tm.get(entity);
+                Vector2 offset = VectorUtils.rotateVector(fc.offset, targetPos.rotation);
+
                 switch (fc.followMode) {
                     case STICKY:
-                        TransformComponent tc = tm.get(entity);
-                        Vector2 offset = VectorUtils.rotateVector(fc.offset, targetPos.rotation);
                         tc.position.set(targetPos.position.x + offset.x,
                                 targetPos.position.y + offset.y,
                                 tc.position.z);
-                        tc.setOpacity(targetPos.tint.a);
-                        tc.setRotation(fc.baseRotation + targetPos.rotation);
+                        if(fc.shouldMatchOpacity) {
+                            tc.setOpacity(targetPos.tint.a);
+                        }
+                        if(fc.shouldMatchParentRotation) {
+                            tc.setRotation(fc.baseRotation + targetPos.rotation);
+                        }
                         break;
                     case MOVETO:
-
+                        processMoveToAdjustment(deltaTime, fc, targetPos, tc, offset);
+                        break;
+                    case MOVETOSTICKY:
+                        processMoveToAdjustment(deltaTime, fc, targetPos, tc, offset);
+                        if(tc.position.x == (targetPos.position.x + offset.x) &&
+                           tc.position.y == (targetPos.position.y + offset.y)){
+                            fc.setMode(FollowMode.STICKY);
+                        }
                         break;
                     default:
                         break;
                 }
             }
+        }
+    }
+
+    private void processMoveToAdjustment(float deltaTime, FollowerComponent fc, TransformComponent targetPos, TransformComponent tc, Vector2 offset) {
+        moveToAdjustment.set(targetPos.position.x + offset.x, targetPos.position.y + offset.y);
+        moveToAdjustment.sub(tc.position.x, tc.position.y)
+                        .nor()
+                        .scl(fc.followSpeed * deltaTime);
+
+        float newX = tc.position.x + moveToAdjustment.x;
+        float newY = tc.position.y + moveToAdjustment.y;
+
+        if(tc.position.x < targetPos.position.x + offset.x){
+            newX = MathUtils.clamp(newX, tc.position.x, targetPos.position.x + offset.x);
+        }else{
+            newX = MathUtils.clamp(newX, targetPos.position.x + offset.x, tc.position.x);
+        }
+
+        if(tc.position.y < targetPos.position.y + offset.y){
+            newY = MathUtils.clamp(newY, tc.position.y, targetPos.position.y + offset.y);
+        }else{
+            newY = MathUtils.clamp(newY, targetPos.position.y + offset.y, tc.position.y);
+        }
+
+        tc.position.set(newX, newY, tc.position.z);
+
+        if(fc.shouldMatchOpacity) {
+            tc.setOpacity(targetPos.tint.a);
+        }
+
+        if(fc.shouldMatchParentRotation) {
+            tc.setRotation(fc.baseRotation + targetPos.rotation);
         }
     }
 }